Address FIX-77 QA round 2
